requirements.txt
aiohappyeyeballs==2.6.1
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-lru==2.0.5
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.8.0
dotenv==0.9.9
fonttools==4.56.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.3.0
outcome==1.3.0.post0
pillow==11.1.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydyf==0.11.0
PyMySQL==1.1.1
pyphen==0.17.2
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
qq-botpy==1.2.1
requests==2.32.3
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
tabulate==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
zopfli==0.2.3.post1
pandas~=2.2.3
openpyxl~=3.1.5
aiofiles~=24.1.0